requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
Pillow>=10.0.0
